+++ core/pioreactor/web/static/static/Edit-QVFh3gE2.js
@@ -1,1 +1,0 @@
-import{c as t,j as o}from"./index-D8fOP7l4.js";const a=t(o.jsx("path",{d:"M3 17.25V21h3.75L17.81 9.94l-3.75-3.75zM20.71 7.04c.39-.39.39-1.02 0-1.41l-2.34-2.34a.996.996 0 0 0-1.41 0l-1.83 1.83 3.75 3.75z"}));export{a as E};