+++ core/pioreactor/web/static/static/MoreVert-Cte5dSKf.js
@@ -1,1 +1,0 @@
-import{c as s,j as c}from"./index-D8fOP7l4.js";const t=s(c.jsx("path",{d:"M12 8c1.1 0 2-.9 2-2s-.9-2-2-2-2 .9-2 2 .9 2 2 2m0 2c-1.1 0-2 .9-2 2s.9 2 2 2 2-.9 2-2-.9-2-2-2m0 6c-1.1 0-2 .9-2 2s.9 2 2 2 2-.9 2-2-.9-2-2-2"}));export{t as M};