+++ core/pioreactor/web/static/static/getValidReactChildren-BXSmac_S.js
@@ -1,0 +1,1 @@
+import{r}from"./index-1U6lI47L.js";function a(t){return r.Children.toArray(t).filter(e=>r.isValidElement(e))}export{a as g};